+++ changelog.md
@@ -4,6 +4,7 @@
 
 ### *2014.10.13* - 1.0.3
 * **Fix** - Comment date link color.
+* **Fix** - Comment reply link color.
 
 ### *2014.10.12* - 1.0.2
 * **Fix** - Upsell display on cart when using full width template.