+++ changelog.md
@@ -2,9 +2,10 @@
 
 =====
 
-### *2014.10.13* - 1.0.3
+### *2014.10.15* - 1.0.3
 * **Fix** - Comment date link color.
 * **Fix** - Comment reply link color.
+* **Tweak** - Sanitize layout setting.
 
 ### *2014.10.12* - 1.0.2
 * **Fix** - Upsell display on cart when using full width template.